requirements.txt
acvl_utils==0.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
batchgenerators==0.25
batchgeneratorsv2==0.2
certifi==2024.7.4
charset-normalizer==2.0.12
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
dataclasses==0.6
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dicom2nifti==2.4.11
dill==0.3.8
dynamic_network_architectures==0.3.1
einops==0.8.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
graphviz==0.20.3
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1680185618122/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257377578/work
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mkl==2021.4.0
mpmath==1.3.0
multiprocess==0.70.16
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
nibabel==5.2.1
nnunetv2==2.5.1
numpy==1.26.4
opencv-python==4.10.0.84
p_tqdm==1.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pathos==0.3.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydicom==2.4.4
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-gdcm==3.0.24.1
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
requests==2.27.1
rt-utils==1.2.7
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
SimpleITK==2.3.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.13.1
tbb==2021.13.0
threadpoolctl==3.5.0
tifffile==2024.7.24
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tornado @ file:///D:/bld/tornado_1656937934674/work
TotalSegmentator==2.4.0
tqdm==4.66.4
traceback2==1.4.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
ultralytics==8.2.60
ultralytics-thop==2.0.0
unittest2==1.1.0
urllib3==1.26.19
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
xvfbwrapper==0.2.9
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
